jump search

+++ 0x1E-search_algorithms/100-jump.c
@@ -12,13 +12,13 @@
   */
 int jump_search(int *array, size_t size, int value)
 {
-	size_t i, jump, step;
+	int i, jump, step;
 
 	if (array == NULL || size == 0)
 		return (-1);
 
-	step = sqrt(size);
-	for (i = jump = 0; jump < size && array[jump] < value;)
+	step = sqrt((int)size);
+	for (i = jump = 0; jump < (int)size && array[jump] < value;)
 	{
 		printf("Value checked array[%ld] = [%d]\n", jump, array[jump]);
 		i = jump;
@@ -27,7 +27,7 @@ int jump_search(int *array, size_t size, int value)
 
 	printf("Value found between indexes [%ld] and [%ld]\n", i, jump);
 
-	jump = jump < size - 1 ? jump : size - 1;
+	jump = jump < (int)size - 1 ? jump : (int)size - 1;
 	for (; i < jump && array[i] < value; i++)
 		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 	printf("Value checked array[%ld] = [%d]\n", i, array[i]);
